create Exceptions and create data verification methods

+++ src/main/java/com/aliakseikul/storenew/service/impl/UserServiceImpl.java
@@ -1,8 +1,8 @@
 package com.aliakseikul.storenew.service.impl;
 
 import com.aliakseikul.storenew.entity.User;
-import com.aliakseikul.storenew.entity.enums.ProductBrand;
-import com.aliakseikul.storenew.entity.enums.ProductCategory;
+
+import com.aliakseikul.storenew.exeption.checkMethods.Check;
 import com.aliakseikul.storenew.exeption.exeptions.*;
 import com.aliakseikul.storenew.exeption.message.ErrorMessage;
 import com.aliakseikul.storenew.repository.UserRepository;
@@ -11,9 +11,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.UUID;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static com.aliakseikul.storenew.exeption.checkMethods.Check.checkNumber;
+import static com.aliakseikul.storenew.exeption.checkMethods.Check.valueNullOrEmpty;
 
 @Service
 @RequiredArgsConstructor
@@ -23,19 +26,31 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findById(String id) {
-        checkIdLength(id);
-        return userRepository.findById(UUID.fromString(id)).orElse(null);
+        if (Check.checkIdLength(id)) {
+            throw new UserNotFoundException(ErrorMessage.WRONG_ID_LENGTH);
+        }
+        return userRepository.findById(UUID.fromString(id))
+                .orElseThrow(() -> new UserNotFoundException(ErrorMessage.USER_NOT_FOUND));
     }
 
     @Override
     public User addUser(User user) {
+        if (user == null) {
+            throw new IllegalArgumentException(ErrorMessage.NULL_OR_EMPTY);
+        }
         return userRepository.save(user);
     }
 
     @Override
     public ResponseEntity<String> updateProductParamById(String userId, String property, String value) {
-        checkId(userId);
+        if (checkId(userId)) {
+            throw new UserNotFoundException(ErrorMessage.USER_NOT_FOUND);
+        }
+        if (valueNullOrEmpty(property) || valueNullOrEmpty(value)) {
+            throw new StringIsNullExceptions(ErrorMessage.NULL_OR_EMPTY);
+        }
         UUID userUuid = UUID.fromString(userId);
+
         String responseMessage;
         switch (property.toLowerCase()) {
             case "name":
@@ -47,10 +62,16 @@ public class UserServiceImpl implements UserService {
                 responseMessage = "set new lastname " + value;
                 break;
             case "email":
+                if (checkEmail(value)) {
+                    throw new NumberExceptions(ErrorMessage.WRONG_EMAIL);
+                }
                 userRepository.changeEmailUserById(userUuid, value);
                 responseMessage = "set new email " + value;
                 break;
             case "phonenumber":
+                if (checkNumber(value)) {
+                    throw new EmailExceptions(ErrorMessage.NUMBER_ERROR);
+                }
                 userRepository.changePhoneNumberUserById(userUuid, value);
                 responseMessage = "set new phone number " + value;
                 break;
@@ -62,26 +83,27 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUserById(String userId) {
-        checkId(userId);
+        if (checkId(userId)) {
+            throw new UserNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND);
+        }
         userRepository.deleteById(UUID.fromString(userId));
     }
 
-    private boolean valueNullOrEmpty(String value) {
-        return value == null || value.isEmpty();
+    private boolean checkEmail(String value) {
+        if (valueNullOrEmpty(value)) {
+            throw new EmailExceptions(ErrorMessage.NULL_OR_EMPTY);
+        }
+        String patter = "\\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Z|a-z]{2,}\\b";
+
+        Pattern pattern = Pattern.compile(patter);
+        Matcher matcher = pattern.matcher(value);
+        return matcher.matches();
     }
 
-    private void checkIdLength(String productId) {
-        if (valueNullOrEmpty(productId)) {
-            throw new ProductNotFoundException(ErrorMessage.NULL_OR_EMPTY);
+    private boolean checkId(String userId) {
+        if (valueNullOrEmpty(userId)) {
+            throw new UserNotFoundException(ErrorMessage.NULL_OR_EMPTY);
         }
-        if (productId.length() != 36) {
-            throw new ProductNotFoundException(ErrorMessage.WRONG_ID_LENGTH);
-        }
-    }
-
-    private void checkId(String productId) {
-        if (findById(productId) == null) {
-            throw new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND);
-        }
+        return findById(userId) == null;
     }
 }
